requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.0
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.0
fonttools==4.39.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
lap==0.4.0
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.17.0
six==1.16.0
supervision==0.3.1
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
ultralytics==8.0.98
urllib3==1.26.15
kafka-python==2.0.2
moviepy
jsonpickle
onnx2pytorch
pyspark
prometheus_client
